refactor(app): address PR review feedback
- Delete plan document (not for commit)
- Extract usePopoverAnimation hook to separate file
- Extract ConvertToTransferMenuItem to separate component file
